+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 1.0.3
+
+* Add `isMuted` getter to determine mute status.
+* Update `notifyListeners` to return true iff unmuted call was made.
+
+## 1.0.1
+
+* Discontinued - renamed to "muteable_notifiers".
+
 ## 1.0.0
 
 * Initial commit of MutableChangeNotifier and MutableValueNotifier.